Finally fix node not switching to candidate.
Turns out I only shutdown the gRPC service of a node on shutdown, not
the state machine. The leader was not accessible externally but it was
still able to communicate with the followers. The followers don't need
the leader gRPC endpoint to be available to know that it's available.
They just need to know that the leader is still sending append entries
RPC regularly.